requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0
scikit-learn>=1.3.0
pyyaml>=6.0
pillow>=9.5.0


transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0
peft>=0.6.0
accelerate>=0.24.0


